test: consolidate vector tests into functions.test.js
- Move comprehensive vector function tests to functions.test.js
- Add L2NORMALIZE tests (was missing)
- Add semantic search tests (ORDER BY similarity, dynamic embedding)
- Remove separate vector.test.js
